--- src/pages/lecture/styled.ts
@@ -1,11 +1,0 @@
-import { chakra, Box } from '@chakra-ui/react';
-
-export const Container = chakra(Box, {
-  baseStyle: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: '100%',
-  },
-});
