+++ src/pages/roadmap/index.tsx
@@ -1,14 +1,23 @@
 import React from 'react';
+import { chakra, Box } from '@chakra-ui/react';
 
 import RoadmapSvg from '#/components/UI/atomics/RoadmapSVG';
 import RoadmapFieldContainer from '#/components/UI/organisms/RoadmapFieldContainer';
-import * as Styled from './styled';
+
+export const RoadmapContainer = chakra(Box, {
+  baseStyle: {
+    position: 'relative',
+    top: '0',
+    width: '100%',
+    height: '89vh',
+  },
+});
 
 const RoadMapPage = () => (
-  <Styled.RoadmapContainer>
+  <RoadmapContainer>
     <RoadmapSvg />
     <RoadmapFieldContainer />
-  </Styled.RoadmapContainer>
+  </RoadmapContainer>
 );
 
 export default RoadMapPage;
