--- src/pages/roadmap/styled.ts
@@ -1,10 +1,0 @@
-import { chakra, Box } from '@chakra-ui/react';
-
-export const RoadmapContainer = chakra(Box, {
-  baseStyle: {
-    position: 'relative',
-    top: '0',
-    width: '100%',
-    height: '89vh',
-  },
-});
